Add color data & converting & decorating functions
All this had to be added at the same time, since get256color() needs
palette256 to work, and fmt() needs get256color() to work. palette256 is
a named vector of numbers from 0 to 255 whose names are the Xterm names
for those colors. get256color() converts a number from 0 to 255 or
anything R recognizes as a color to an Xterm color. fmt() is the main
function and decorates console text, setting the text color, background
color, bold, italic, and/or underline.