Improve documentation and increment version
When creating an image to illustrate fmt(), I cleaned up the example
code a little but forgot to update the example code, so this fixes that.
I had also left the documentation for palette256 a little unfinished,
and this commit fixes that too. Finally, with all that done, I'm ready
for the initial version of the package.